requirements.txt
# Core dependencies for AppDynamics data collection and SRE analytics
requests>=2.31.0
PyYAML>=6.0.1
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2

# Data visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data processing and analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Configuration and logging
python-dotenv>=1.0.0
colorlog>=6.7.0

# Time series analysis
statsmodels>=0.14.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Report generation (HTML/PDF)
jinja2>=3.1.0
weasyprint>=59.0
reportlab>=4.0.0

# LLM integration for intelligent analysis
openai>=1.0.0
anthropic>=0.18.0

# Web UI for report generation
flask>=2.3.2

# Optional: For advanced analytics
# jupyter>=1.0.0
# ipykernel>=6.25.0